Move functional component analysis into separate file
